requirements.txt
websocket-client==1.8.0
rel==0.4.9.19
requests==2.32.3